requirements.txt
djangorestframework==3.12.4
requests==2.26.0
psycopg2-binary==2.9.1